requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pdfplumber>=0.10.0
fpdf2>=2.7.0
requests>=2.28.0
anthropic>=0.25.0
python-dotenv
